+++ packages/components/src/List/Toolbar/ColumnChooserButton/ColumnChooser/ColumnChooserTable/index.js
@@ -1,0 +1,3 @@
+import Component from './ColumnChooserTable.component';
+
+export default Component;
